# public/openapi.yaml
openapi: 3.0.3
info:
  title: Fly2Any API
  description: |
    Comprehensive flight booking and management API with AI-powered search optimization,
    multi-source aggregation (Amadeus, Duffel), and intelligent caching.

    ## Features
    - Multi-airport and flexible date search
    - Real-time price confirmation
    - Seat selection and branded fares
    - Payment processing via Stripe
    - Booking management and modifications
    - Price alerts and saved searches

    ## Base URLs
    - Production: https://fly2any.com/api
    - Development: http://localhost:3000/api
  version: 1.0.0
  contact:
    name: Fly2Any API Support
    email: api-support@fly2any.com
    url: https://fly2any.com/docs
  license:
    name: Proprietary
    url: https://fly2any.com/terms
servers:
  - url: https://fly2any.com/api
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
...
security:
  - sessionAuth: []
paths:
...
  /saved-searches:
    get:
      tags:
        - Saved Searches
      summary: List saved searches
      description: Get all saved searches for authenticated user
      operationId: listSavedSearches
      responses:
        '200':
          description: Saved searches retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  searches:
                    type: array
                    items:
                      $ref: '#/components/schemas/SavedSearch'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Saved Searches
      summary: Create saved search
      description: Save a flight search for later use
      operationId: createSavedSearch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SavedSearchInput'
      responses:
        '201':
          description: Search saved
          content:
            application/json:
              schema:
                type: object
                properties:
                  search:
                    $ref: '#/components/schemas/SavedSearch'
                  created:
                    type: boolean
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

    put:
      tags:
        - Saved Searches
      summary: Update saved search
      description: Update an existing saved search
      operationId: updateSavedSearch
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                filters:
                  type: object
      responses:
        '200':
          description: Search updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  search:
                    $ref: '#/components/schemas/SavedSearch'
                  updated:
                    type: boolean
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Saved Searches
      summary: Delete saved search
      description: Delete a saved search
      operationId: deleteSavedSearch
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Search deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  deleted:
                    type: boolean
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /price-alerts:
    get:
      tags:
        - Price Alerts
      summary: List price alerts
      description: Get all price alerts for authenticated user
      operationId: listPriceAlerts
      parameters:
        - name: active
          in: query
          schema:
            type: boolean
          description: Filter active alerts only
      responses:
        '200':
          description: Price alerts retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  alerts:
                    type: array
                    items:
                      $ref: '#/components/schemas/PriceAlert'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Price Alerts
      summary: Create price alert
      description: Create a new price alert for a route
      operationId: createPriceAlert
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PriceAlertInput'
      responses:
        '201':
          description: Alert created
          content:
            application/json:
              schema:
                type: object
                properties:
                  alert:
                    $ref: '#/components/schemas/PriceAlert'
                  created:
                    type: boolean
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

    patch:
      tags:
        - Price Alerts
      summary: Update price alert
      description: Update an existing price alert
      operationId: updatePriceAlert
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                active:
                  type: boolean
                targetPrice:
                  type: number
                currentPrice:
                  type: number
      responses:
        '200':
          description: Alert updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  alert:
                    $ref: '#/components/schemas/PriceAlert'
                  updated:
                    type: boolean
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Price Alerts
      summary: Delete price alert
      description: Delete a price alert
      operationId: deletePriceAlert
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Alert deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  deleted:
                    type: boolean
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: Session-based authentication via NextAuth.js
  schemas:
...
    SavedSearch:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        name:
          type: string
        origin:
          type: string
        destination:
          type: string
        departDate:
          type: string
          format: date
        returnDate:
          type: string
          format: date
        adults:
          type: integer
        children:
          type: integer
        infants:
          type: integer
        cabinClass:
          type: string
        filters:
          type: object
        searchCount:
          type: integer
        lastSearched:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    SavedSearchInput:
      type: object
      required:
        - name
        - origin
        - destination
        - departDate
        - adults
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        origin:
          type: string
          minLength: 3
          maxLength: 3
        destination:
          type: string
          minLength: 3
          maxLength: 3
        departDate:
          type: string
          format: date
        returnDate:
          type: string
          format: date
        adults:
          type: integer
          minimum: 1
          maximum: 9
        children:
          type: integer
          minimum: 0
          maximum: 9
          default: 0
        infants:
          type: integer
          minimum: 0
          maximum: 9
          default: 0
        cabinClass:
          type: string
          enum: [economy, premium_economy, business, first]
          default: economy
        filters:
          type: object
    PriceAlert:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        origin:
          type: string
        destination:
          type: string
        departDate:
          type: string
          format: date
        returnDate:
          type: string
          format: date
        currentPrice:
          type: number
        targetPrice:
          type: number
        currency:
          type: string
        active:
          type: boolean
        triggered:
          type: boolean
        triggeredAt:
          type: string
          format: date-time
        lastChecked:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    PriceAlertInput:
      type: object
      required:
        - origin
        - destination
        - departDate
        - currentPrice
        - targetPrice
      properties:
        origin:
          type: string
          minLength: 3
          maxLength: 3
        destination:
          type: string
          minLength: 3
          maxLength: 3
        departDate:
          type: string
          format: date
        returnDate:
          type: string
          format: date
        currentPrice:
          type: number
          minimum: 0
        targetPrice:
          type: number
          minimum: 0
        currency:
          type: string
          default: USD
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          default: false
        error:
          oneOf:
            - type: string
            - type: object
              properties:
                code:
                  type: string
                message:
                  type: string
                details:
                  type: object
        meta:
          type: object
          properties:
            timestamp:
              type: string
              format: date-time
  responses:
    BadRequest:
      description: Bad request - invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...